🏗️ Refactor App.tsx into modular components and hooks
Major refactoring following CodeRabbit's recommendations:
- Extracted custom hooks for logic separation (useSessionManager, useMobileSessionSync, etc.)
- Created dedicated session components (SessionManager, MobileSessionProcessor)
- Added Zustand store for centralized state management
- Fixed TypeScript errors and build issues
- Added debug logging to trace session creation flow
- Removed ConvexDemo component from desktop app

App.tsx reduced from 942 to 142 lines (85% reduction)
